攻撃回数をIntからDoubleに変更
装備品の攻撃回数ボーナスが反映されないバグを修正。
10倍スケール方式を廃止し、Doubleで直接保持するように変更。

- ItemDefinition.CombatBonuses.attackCount: Int → Double
- CharacterValues.Combat.attackCount: Int → Double
- ItemMaster.jsonのattackCount値を1/10に変換
- SQLiteスキーマをREAL型に変更
- 装備ボーナス適用後に最低1回を保証

+++ Epika/Application/GameRuntime/Battle/BattleTurnEngine.EnemySpecialSkill.swift
@@ -504,7 +504,7 @@ extension BattleTurnEngine {
         case .accuracy:
             actor.snapshot.hitRate = Int(Double(actor.snapshot.hitRate) * multiplier)
         case .attackCount:
-            actor.snapshot.attackCount = Int(Double(actor.snapshot.attackCount) * multiplier)
+            actor.snapshot.attackCount = actor.snapshot.attackCount * multiplier
         }
     }
 }
